+++ CHANGES.rst
@@ -38,6 +38,11 @@ astropy.coordinates
 - Added a ``concatenate_representations`` function to combine coordinate
   representation data and any associated differentials. [#7922]
 
+- ``BaseCoordinateFrame`` will now check for a method named
+  ``_astropy_repr_in_frame`` when constructing the string forms of attributes.
+  Allowing any class to control how ``BaseCoordinateFrame`` represents it when
+  it is an attribute of a frame. [#7745]
+
 - Some rarely-changed attributes of frame classes are now cached, resulting in
   speedups (up to 50% in some cases) when creating new scalar frame or
   ``SkyCoord`` objects. [#7949]